changes socketio for pusher because of vercel

+++ actions/publicMessages/delete.ts
@@ -2,7 +2,27 @@
 import { prisma } from "@/server/db";
 
 export async function deletePublicMessage(id: string) {
-    return prisma.message.delete({
-        where: { id, type: 'public' }
-    });
+    try {
+        const message = await prisma.message.findUnique({
+            where: {
+                id,
+                type: { in: ['public', 'text', 'gif'] }
+            }
+        });
+
+        if (!message) {
+            console.log(`Message with id ${id} not found`);
+            return null;
+        }
+
+        return await prisma.message.delete({
+            where: {
+                id,
+                type: { in: ['public', 'text', 'gif'] }
+            }
+        });
+    } catch (error) {
+        console.error("Error deleting message:", error);
+        throw error;
+    }
 }
